+++ src/vectors/vectors.h
@@ -10,15 +10,15 @@
 
 typedef struct s_vector2
 {
-	float	u;
-	float	v;
+	double	u;
+	double	v;
 }	t_vector2;
 
 typedef struct s_vector3
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vector3;
 
 typedef t_vector3	t_point3;
@@ -27,31 +27,31 @@ typedef t_vector3	t_point3;
 //	@param u The u component
 //	@param v The v component
 //	@return A new 2D vector
-t_vector2	vector2(float u, float v);
+t_vector2	vector2(double u, double v);
 
 //	vector3: Create a new 3D vector
 //	@param x The x component
 //	@param y The y component
 //	@param z The z component
 //	@return A new 3D vector
-t_vector3	vector3(float x, float y, float z);
+t_vector3	vector3(double x, double y, double z);
 
 //	point3: Create a new 3D point
 //	@param x The x component
 //	@param y The y component
 //	@param z The z component
 //	@return A new 3D point
-t_point3	point3(float x, float y, float z);
+t_point3	point3(double x, double y, double z);
 
 //	vlength: Get the length of a vector
 //	@param v The vector
 //	@return The length of the vector
-float		vlength(t_vector3 v);
+double		vlength(t_vector3 v);
 
 //	vlength2: Get the squared length of a vector
 //	@param v The vector
 //	@return The squared length of the vector
-float		vlength2(t_vector3 v);
+double		vlength2(t_vector3 v);
 
 //	vadd: Add two vectors
 //	@param v The first vector
@@ -69,19 +69,19 @@ t_vector3	vsub(t_vector3 v, t_vector3 u);
 //	@param v The vector
 //	@param scalar The scalar
 //	@return The product of the vector and the scalar
-t_vector3	vmul(t_vector3 v, float scalar);
+t_vector3	vmul(t_vector3 v, double scalar);
 
 //	vdiv: Divide a vector by a scalar
 //	@param v The vector
 //	@param scalar The scalar
 //	@return The quotient of the vector and the scalar
-t_vector3	vdiv(t_vector3 v, float scalar);
+t_vector3	vdiv(t_vector3 v, double scalar);
 
 //	vdot: Get the dot product of two vectors
 //	@param v The first vector
 //	@param u The second vector
 //	@return The dot product of the two vectors
-float		vdot(t_vector3 v, t_vector3 u);
+double		vdot(t_vector3 v, t_vector3 u);
 
 //	vcross: Get the cross product of two vectors
 //	@param v The first vector
