+++ src/vectors/vectors_instance.c
@@ -1,6 +1,6 @@
 #include "vectors.h"
 
-t_vector3	vector3(float x, float y, float z)
+t_vector3	vector3(double x, double y, double z)
 {
 	t_vector3	vector;
 
@@ -10,12 +10,12 @@ t_vector3	vector3(float x, float y, float z)
 	return (vector);
 }
 
-t_point3	point3(float x, float y, float z)
+t_point3	point3(double x, double y, double z)
 {
 	return (vector3(x, y, z));
 }
 
-t_vector2	vector2(float u, float v)
+t_vector2	vector2(double u, double v)
 {
 	t_vector2	vector;
